refactor(audit): A3 stage 1 — domain purity for AppInfo + export/share use cases (#45/#18/#19)
Low-risk tier of the A3 domain-purity refactor (behavior-preserving relocation):

- AppInfo (#45): moved the framework->domain companion mapToAppInfo() (+ its getInstallerPackageName
  helper) and all android.content.pm / android.os imports out of domain/model/AppInfo.kt into a new
  data-layer mapper (data/repository/AppInfoMapper.kt). AppInfo is now a pure Kotlin data class (kept
  @Immutable Compose-stability hint, @Serializable, and pure helpers). enabledState default inlined
  to 0 (== the framework's COMPONENT_ENABLED_STATE_DEFAULT). AppRepositoryImpl's 3 call sites updated.
- ExportAppUseCase / ShareAppUseCase (#18/#19): moved the MediaStore / SAF / DocumentFile /
  FileProvider / ContentResolver / Environment I/O behind a new pure-Kotlin domain port
  (domain/repository/AppBundleFileStore) with a data impl (AppBundleFileStoreImpl,
  @Single(binds=[AppBundleFileStore::class]) auto-discovered by @ComponentScan). ExportAppUseCase is
  now fully android-free; ShareAppUseCase keeps only a thin android.net.Uri residual on its return
  (the FileProvider content-uri round-trips through String). Public use-case signatures unchanged, so
  consumers (ExportBottomSheet, MainViewModel) are untouched. Write/overwrite/SAF/Downloads/label
  behavior moved verbatim.

Build: assembleFossDebug green.
