+++ app/src/main/java/com/valhalla/thor/domain/usecase/ShareAppUseCase.kt
@@ -1,10 +1,8 @@
 package com.valhalla.thor.domain.usecase
 
-import android.content.Context
 import android.net.Uri
-import androidx.core.content.FileProvider
-import com.valhalla.thor.BuildConfig
 import com.valhalla.thor.domain.model.AppInfo
+import com.valhalla.thor.domain.repository.AppBundleFileStore
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import org.koin.core.annotation.Factory
@@ -12,13 +10,15 @@ import org.koin.core.annotation.Named
 
 @Factory
 class ShareAppUseCase(
-    private val context: Context,
     private val bundleBuilder: AppBundleBuilder,
+    private val fileStore: AppBundleFileStore,
     @Named("io") private val ioDispatcher: CoroutineDispatcher
 ) {
     suspend operator fun invoke(appInfo: AppInfo): Result<Uri> = withContext(ioDispatcher) {
         bundleBuilder.build(appInfo).mapCatching { file ->
-            FileProvider.getUriForFile(context, "${BuildConfig.APPLICATION_ID}.provider", file)
+            // The FileProvider content-uri bulk lives in the data layer; the use case keeps only
+            // a thin android.net.Uri residual for its return type (MainViewModel's stable contract).
+            Uri.parse(fileStore.shareUri(file))
         }
     }
 }
